refactor, feat: 패키지 위치 변경 및 세부 repository 추가
